refactor(vm): separate kvm memory slots from GuestMemoryMmap
Remove the assumption that GuestMemoryMmap is the same as kvm memory
slots.

Also, make the `register_memory_region` operation internally mutable
using a Mutex so that we don't need a borrow on the `Vm` structure to
call it (we will need it later as we dynamically change it).

HACK: the dirty bitmap page tracking was hacked together and not tested.
It now indexes on the gpa rather than slot number and is able to handle
the case where all kvm regions are fully included in a GuestMemoryMmap
region.
I think the approach is sound but need to double check.

+++ src/vmm/src/lib.rs
@@ -261,7 +261,7 @@ pub enum VmmError {
 }
 
 /// Shorthand type for KVM dirty page bitmap.
-pub type DirtyBitmap = HashMap<u32, Vec<u64>>;
+pub type DirtyBitmap = HashMap<u64, Vec<u64>>;
 
 /// Returns the size of guest memory, in MiB.
 pub(crate) fn mem_size_mib(guest_memory: &GuestMemoryMmap) -> u64 {
